requirements.txt
Django==2.2
django-cors-headers==3.5.0
django-restframework==0.0.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
